http/extra/django_api/requirements.txt
certifi==2024.8.30
chardet==5.2.0
Django==5.1.4
djangorestframework==3.15.2
idna==3.10
pytz==2024.2
requests==2.32.3
urllib3==2.2.3
